HDD support, based on AO486
