+++ MiSTer.vcxproj
@@ -72,10 +72,6 @@
     <ClCompile Include="lib\miniz\miniz_tdef.c" />
     <ClCompile Include="lib\miniz\miniz_tinfl.c" />
     <ClCompile Include="lib\miniz\miniz_zip.c" />
-    <ClCompile Include="lib\serial_server\library\Checksum.cpp" />
-    <ClCompile Include="lib\serial_server\library\Image.cpp" />
-    <ClCompile Include="lib\serial_server\library\Process.cpp" />
-    <ClCompile Include="lib\serial_server\library\Serial.cpp" />
     <ClCompile Include="main.cpp" />
     <ClCompile Include="menu.cpp" />
     <ClCompile Include="offload.cpp" />
@@ -104,7 +100,6 @@
     <ClCompile Include="support\pcecd\pcecd.cpp" />
     <ClCompile Include="support\pcecd\pcecdd.cpp" />
     <ClCompile Include="support\pcecd\seektime.cpp" />
-    <ClCompile Include="support\pcxt\pcxt.cpp" />
     <ClCompile Include="support\psx\psx.cpp" />
     <ClCompile Include="support\sharpmz\sharpmz.cpp" />
     <ClCompile Include="support\snes\snes.cpp" />
@@ -150,10 +145,6 @@
     <ClInclude Include="lib\miniz\miniz_tdef.h" />
     <ClInclude Include="lib\miniz\miniz_tinfl.h" />
     <ClInclude Include="lib\miniz\miniz_zip.h" />
-    <ClInclude Include="lib\serial_server\library\FlatImage.h" />
-    <ClInclude Include="lib\serial_server\library\Library.h" />
-    <ClInclude Include="lib\serial_server\library\LinuxFile.h" />
-    <ClInclude Include="lib\serial_server\library\LinuxSerial.h" />
     <ClInclude Include="logo.h" />
     <ClInclude Include="menu.h" />
     <ClInclude Include="offload.h" />
@@ -181,7 +172,6 @@
     <ClInclude Include="support\minimig\minimig_share.h" />
     <ClInclude Include="support\neogeo\neogeo_loader.h" />
     <ClInclude Include="support\pcecd\pcecd.h" />
-    <ClInclude Include="support\pcxt\pcxt.h" />
     <ClInclude Include="support\psx\mcdheader.h" />
     <ClInclude Include="support\psx\psx.h" />
     <ClInclude Include="support\sharpmz\sharpmz.h" />
