+++ support/x86/x86.cpp
@@ -51,6 +51,163 @@
 
 #define IOWR(base, reg, value) x86_dma_set((base) + (reg), value)
 
+bool pcxt = false;
+unsigned int hdd_table[128][3] = {
+	{  306,  4, 17 },		/* 0 - 7 */
+	{  615,  2, 17 },
+	{  306,  4, 26 },
+	{ 1024,  2, 17 },
+	{  697,  3, 17 },
+	{  306,  8, 17 },
+	{  614,  4, 17 },
+	{  615,  4, 17 },
+
+	{  670,  4, 17 },		/* 8 - 15 */
+	{  697,  4, 17 },
+	{  987,  3, 17 },
+	{  820,  4, 17 },
+	{  670,  5, 17 },
+	{  697,  5, 17 },
+	{  733,  5, 17 },
+	{  615,  6, 17 },
+
+	{  462,  8, 17 },		/* 016-023 */
+	{  306,  8, 26 },
+	{  615,  4, 26 },
+	{ 1024,  4, 17 },
+	{  855,  5, 17 },
+	{  925,  5, 17 },
+	{  932,  5, 17 },
+	{ 1024,  2, 40 },
+
+	{  809,  6, 17 },		/* 024-031 */
+	{  976,  5, 17 },
+	{  977,  5, 17 },
+	{  698,  7, 17 },
+	{  699,  7, 17 },
+	{  981,  5, 17 },
+	{  615,  8, 17 },
+	{  989,  5, 17 },
+
+	{  820,  4, 26 },		/* 032-039 */
+	{ 1024,  5, 17 },
+	{  733,  7, 17 },
+	{  754,  7, 17 },
+	{  733,  5, 26 },
+	{  940,  6, 17 },
+	{  615,  6, 26 },
+	{  462,  8, 26 },
+
+	{  830,  7, 17 },		/* 040-047 */
+	{  855,  7, 17 },
+	{  751,  8, 17 },
+	{ 1024,  4, 26 },
+	{  918,  7, 17 },
+	{  925,  7, 17 },
+	{  855,  5, 26 },
+	{  977,  7, 17 },
+
+	{  987,  7, 17 },		/* 048-055 */
+	{ 1024,  7, 17 },
+	{  823,  4, 38 },
+	{  925,  8, 17 },
+	{  809,  6, 26 },
+	{  976,  5, 26 },
+	{  977,  5, 26 },
+	{  698,  7, 26 },
+
+	{  699,  7, 26 },		/* 056-063 */
+	{  940,  8, 17 },
+	{  615,  8, 26 },
+	{ 1024,  5, 26 },
+	{  733,  7, 26 },
+	{ 1024,  8, 17 },
+	{  823, 10, 17 },
+	{  754, 11, 17 },
+
+	{  830, 10, 17 },		/* 064-071 */
+	{  925,  9, 17 },
+	{ 1224,  7, 17 },
+	{  940,  6, 26 },
+	{  855,  7, 26 },
+	{  751,  8, 26 },
+	{ 1024,  9, 17 },
+	{  965, 10, 17 },
+
+	{  969,  5, 34 },		/* 072-079 */
+	{  980, 10, 17 },
+	{  960,  5, 35 },
+	{  918, 11, 17 },
+	{ 1024, 10, 17 },
+	{  977,  7, 26 },
+	{ 1024,  7, 26 },
+	{ 1024, 11, 17 },
+
+	{  940,  8, 26 },		/* 080-087 */
+	{  776,  8, 33 },
+	{  755, 16, 17 },
+	{ 1024, 12, 17 },
+	{ 1024,  8, 26 },
+	{  823, 10, 26 },
+	{  830, 10, 26 },
+	{  925,  9, 26 },
+
+	{  960,  9, 26 },		/* 088-095 */
+	{ 1024, 13, 17 },
+	{ 1224, 11, 17 },
+	{  900, 15, 17 },
+	{  969,  7, 34 },
+	{  917, 15, 17 },
+	{  918, 15, 17 },
+	{ 1524,  4, 39 },
+
+	{ 1024,  9, 26 },		/* 096-103 */
+	{ 1024, 14, 17 },
+	{  965, 10, 26 },
+	{  980, 10, 26 },
+	{ 1020, 15, 17 },
+	{ 1023, 15, 17 },
+	{ 1024, 15, 17 },
+	{ 1024, 16, 17 },
+
+	{ 1224, 15, 17 },		/* 104-111 */
+	{  755, 16, 26 },
+	{  903,  8, 46 },
+	{  984, 10, 34 },
+	{  900, 15, 26 },
+	{  917, 15, 26 },
+	{ 1023, 15, 26 },
+	{  684, 16, 38 },
+
+	{ 1930,  4, 62 },		/* 112-119 */
+	{  967, 16, 31 },
+	{ 1013, 10, 63 },
+	{ 1218, 15, 36 },
+	{  654, 16, 63 },
+	{  659, 16, 63 },
+	{  702, 16, 63 },
+	{ 1002, 13, 63 },
+
+	{  854, 16, 63 },		/* 119-127 */
+	{  987, 16, 63 },
+	{  995, 16, 63 },
+	{ 1024, 16, 63 },
+	{ 1036, 16, 63 },
+	{ 1120, 16, 59 },
+	{ 1054, 16, 63 },
+	{    0,  0,  0 }
+};
+
+struct hddInfo {
+	unsigned long size;
+	unsigned long cylinders;
+	unsigned long heads;
+	unsigned long sectors;
+};
+
+struct hddInfo* FindHDDInfoBySize(unsigned long size);
+struct hddInfo hddInfos[] = { { 0, 0, 0, 0 } };
+
 typedef struct
 {
 	uint32_t ver;
@@ -273,6 +430,9 @@ static void fdd_set(int num, char* filename)
 
 static void hdd_set(int num, char* filename)
 {
+	FILE* fd;
+	long size;
+	struct hddInfo* hdd_fi;
 	int present = 0;
 	int cd = 0;
 
@@ -287,7 +447,34 @@ static void hdd_set(int num, char* filename)
 	}
 
 	if(!present && vhd) present = ide_img_mount(&ide_image[num], filename, 1);
-	ide_img_set(num, present ? &ide_image[num] : 0, cd);
+	if (!cd) {
+		const char* path = getFullPath(filename);
+		fd = fopen(path, "r");
+		if (fd)
+		{
+			fseek(fd, 0L, SEEK_END);
+			size = ftell(fd);
+			if ((hdd_fi = FindHDDInfoBySize(size)))
+			{
+				ide_img_set(num, present ? &ide_image[num] : 0, cd, hdd_fi->sectors, hdd_fi->heads);
+			}
+			else
+			{
+				if (size > 8455200768) // 16383 cylinders * 16 heads * 63 sectors * 512 bytes per sector (Max. CHS)
+				{
+					ide_img_set(num, present ? &ide_image[num] : 0, cd);
+				}
+				else
+				{
+					ide_img_set(num, present ? &ide_image[num] : 0, cd, 63, 16);
+				}
+			}
+		}
+	}
+	else
+	{
+		ide_img_set(num, present ? &ide_image[num] : 0, cd);
+	}
 }
 
 static uint8_t bin2bcd(unsigned val)
@@ -295,14 +482,19 @@ static uint8_t bin2bcd(unsigned val)
 	return ((val / 10) << 4) + (val % 10);
 }
 
-void x86_init()
+void x86_init(bool _is_pcxt)
 {
 	user_io_status_set("[0]", 1);
 
 	const char *home = HomeDir();
 
-	load_bios(user_io_make_filepath(home, "boot0.rom"), 0);
-	load_bios(user_io_make_filepath(home, "boot1.rom"), 1);
+	pcxt = _is_pcxt;
+
+	if (!_is_pcxt)
+	{
+		load_bios(user_io_make_filepath(home, "boot0.rom"), 0);
+		load_bios(user_io_make_filepath(home, "boot1.rom"), 1);
+	}
 
 	uint16_t cfg = ide_check();
 	uint8_t hotswap[4] = {
@@ -580,4 +772,31 @@ const char* x86_get_image_name(int num)
 const char* x86_get_image_path(int num)
 {
 	return config.img_name[num];
+}
+
+struct hddInfo* FindHDDInfoBySize(unsigned long size)
+{
+	struct hddInfo* fi;
+	unsigned long size_chs;
+	bool is_chs = false;
+
+
+	for (int i = 0; i < 127; i++) {
+		size_chs = ((unsigned long)hdd_table[i][0]) * (unsigned long)hdd_table[i][1] * (unsigned long)hdd_table[i][2] * 512;
+		if (size == size_chs)
+		{
+			fi = hddInfos;
+			fi->size = size;
+			fi->cylinders = hdd_table[i][0];
+			fi->heads = hdd_table[i][1];
+			fi->sectors = hdd_table[i][2];
+			is_chs = true;
+			break;
+		}
+	}
+
+	if (!is_chs)
+		fi = NULL;
+
+	return(fi);
 }
